refactor: rename package and use instant for timestamp
